Update the README with a little more information
Add sections for building and other contributing guidelines. We may want to
split the build section out into its own file. We can also add authors as
needed.

This also adds a GitHub action to lint any Markdown files in the repo for
pull requests, so that we can ensure that our docs are in good shape.